remove temporary files
